remote mediator added but now everething without coords

+++ app/src/main/java/ru/netology/nework/dao/PostRemoteKeyDao.kt
@@ -4,15 +4,15 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import ru.netology.nmedia.entity.PostRemoteKeyEntity
+import ru.netology.nework.entity.PostRemoteKeyEntity
 
 @Dao
 interface PostRemoteKeyDao {
     @Query("SELECT max(`key`) FROM PostRemoteKeyEntity")
-    suspend fun max(): Long?
+    suspend fun max(): Int?
 
     @Query("SELECT min(`key`) FROM PostRemoteKeyEntity")
-    suspend fun min(): Long?
+    suspend fun min(): Int?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(postRemoteKeyEntity: PostRemoteKeyEntity)
